requirements.txt
google-generativeai
flask
functions-framework